+++ cantera/trunk/src/transport/TransportBase.cpp
@@ -144,8 +144,8 @@ void Transport::finalize()
 
 //====================================================================================================================
 void Transport::getSpeciesFluxes(size_t ndim, const doublereal* const grad_T,
-                                 int ldx, const doublereal* const grad_X,
-                                 int ldf, doublereal* const fluxes)
+                                 size_t ldx, const doublereal* const grad_X,
+                                 size_t ldf, doublereal* const fluxes)
 {
     err("getSpeciesFluxes");
 }
